+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## Unreleased
+
+- Added `RemoveDefaultTimeToLive` switch parameter to `Set-CosmosDbCollection`
+  to allow removal of a default time to live setting on a collection - fixes [Issue #144](https://github.com/PlagueHO/CosmosDB/issues/144)
+
 ## 2.1.3.528
 
 - Changed `New-CosmosDbStoredProcedure` & `Set-CosmosDbStoredProcedure` to use serialization